+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 2.8
+* Runtime.Core: ME.ECS.Collab minor fixes for Unity 2020
 * Runtime.Essentials: All essential packages has been moved out from the main repo
 * Runtime.Core:
   * Add-ons window added